src/pydantic_ai_agent/pyproject.toml
[project]
name = "pydantic-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai>=1.105.0",
    "pydantic-ai-slim[ag-ui]>=0.8.1",
]
